Add initial database migration and services
This commit includes an initial migration of a Postgres database. We introduced tables such as Authors, Clients, Books, and Genres. Also added entities related to these tables. We specified relationships between tables, such as Books and Authors, and also set primary keys for each table.

Moreover, Postgres database is defined in docker-compose file to allow for quick setup and teardown of the database for development purposes.

The database connection string configuration was added to