"Implemented API error handling and extended author queries"
This commit introduces important changes that enhance API error handling and extend the available author-related operations.

The "Library.Errors.Abstract" project was created for future error abstraction and "BaseAppException" class was formed in it. It will help in consistent handling of exceptions at the application level. Additionally, a generic "EntityNotFoundException" was implemented, this serves to indicate when an entity was not found in the database.

In the author feature, extensive changes have been made to provide more information about an author. A new query called "GetAuthorByIdQuery" was added to fetch details about a particular author by their id including the books related to the author. A new endpoint corresponding to this query was created, named "GetAuthorEndpoint", along with their handler "GetAuthorByIdQueryHandler".

For mapping, methods were added extending the Entity to DTO mappings for Authors and Books. Nullable flag in some project files was disabled for convenience.

Lastly, an application extension for exception handling was implemented. This will ensure that exceptions are caught and proper HTTP Responses are returned, along with detailed error information in JSON.