Refactor Authors API for pagination and filtering
Removed `AuthorWithNumberOfBooksDto` as it was obsolete and added `AuthorWithBooksDto`. Refactored API to support pagination and filtering of authors. Replaced List with IEnumerable for `Items` in `PageResultDto` to make it more flexible. Modified `AuthorExtensions` to convert `PageResult<Author>` to `PageResultDto<AuthorWithBooksDto>`. Updated project dependencies accordingly. These changes were made to improve performance and user experience when dealing with a large number of authors.