Add remove book functionality and refactor author removal
A new RemoveBookCommand has been added to provide the functionality of removing a book. This was necessary to allow users to manage the books in the library. The RemoveAuthorByIdQuery and its respective handler have been renamed to RemoveAuthorByIdCommand and RemoveAuthorByIdCommandHandler respectively, to better reflect their usage as commands not queries. Additionally, the 'Id' field has been added to the BookByAuthorDto class to improve data integrity when mapping entities to this DTO.