Add Get/Edit book functionalities and refine existing.
New files for handling GetBook and EditBook functionalities are created which include necessary DTOs, Commands, Queries, Handlers and Endpoints.

Extra necessary methods added in repositories, mapping extension methods and filters. Also, adjusted calling method in `RemoveBookCommandHandler.cs` and `AuthorRepository.cs` to refine the existing code.

These changes are part of an effort to provide and improve book management features in the library system. It includes handling of retrieving the list of books with pagination, filtering and ability to edit book details. The refined code will provide case-insensitive searching and a more streamlined way of throwing exceptions if book doesn't exist.